+++ src/main/java/ua/edu/sumdu/j2ee/zykov/exception/CustomerNotExistException.java
@@ -1,11 +1,11 @@
 package ua.edu.sumdu.j2ee.zykov.exception;
 
 public class CustomerNotExistException extends NotExistException {
-    public CustomerNotExistException(Long customerId) {
+    public CustomerNotExistException(Integer customerId) {
         super(String.format("Customer %d not exist", customerId));
     }
 
-    public CustomerNotExistException(Long customerId, Throwable cause) {
+    public CustomerNotExistException(Integer customerId, Throwable cause) {
         super(String.format("Customer %d not exist", customerId), cause);
     }
 }
