+++ src/main/java/ua/edu/sumdu/j2ee/zykov/exception/ProductHasOrderNotExistException.java
@@ -1,11 +1,11 @@
 package ua.edu.sumdu.j2ee.zykov.exception;
 
 public class ProductHasOrderNotExistException extends NotExistException {
-    public ProductHasOrderNotExistException(Long productHasOrderId) {
+    public ProductHasOrderNotExistException(Integer productHasOrderId) {
         super(String.format("Product has order %d not exist", productHasOrderId));
     }
 
-    public ProductHasOrderNotExistException(Long productHasOrderId, Throwable cause) {
+    public ProductHasOrderNotExistException(Integer productHasOrderId, Throwable cause) {
         super(String.format("Product has order %d not exist", productHasOrderId), cause);
     }
 }
